Style: BookingResult의 userId 타입을 Long으로 변
- userId 필드 타입을 int에서 Long으로 변경
- userId 생성자 파라미터 및 getter 반환 타입도 함께 Long으로 변경

Why:
1. 테스트에서 사용하는 jdbcTemplate.queryForObject가 Long 타입을 반환
2. 데이터베이스의 id 컬럼이 BIGINT 타입으로 정의되어 있어 Long과 매핑
3. int(최대 약 21억)보다 큰 범위의 Long(약 922경)이 대용량 데이터 처리에 적합

Test:
- testDoubleBookingAfterPayment() 테스트에서 타입 불일치 오류 해결 확인

+++ src/test/java/com/kb/wallet/ticket/dto/BookingResult.java
@@ -1,8 +1,7 @@
 package com.kb.wallet.ticket.dto;
 
 public class BookingResult {
-
-  private final int userId;
+  private final Long userId;
   private final boolean success;
   private final String errorMessage;
   private final long processingTimeNanos;
@@ -10,14 +9,14 @@ public class BookingResult {
 
   public BookingResult(int userId, boolean success, String errorMessage,
       long processingTimeNanos, Long seatId) {
-    this.userId = userId;
+    this.userId = (long) userId;
     this.success = success;
     this.errorMessage = errorMessage;
     this.processingTimeNanos = processingTimeNanos;
     this.seatId = seatId;
   }
 
-  public int getUserId() {
+  public Long getUserId() {
     return userId;
   }
